fetch user via Navbar and logout on hover
