+++ frontend/src/components/Navbar.js
@@ -2,10 +2,11 @@ import React, { useEffect } from 'react';
 import styled from 'styled-components/macro';
 import { StyledNavbar } from './styles/GlobalStyles';
 import { GhostBtn } from './styles/Buttons';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector, useDispatch, batch } from 'react-redux';
 import { Link } from 'react-router-dom';
 import Logo from '../assets/quizzies-logo.png';
 import user from 'reducers/auth';
+import { API_URL } from 'utils/urls';
 
 import { AiOutlineHome } from 'react-icons/ai';
 import { CgProfile } from 'react-icons/cg';
@@ -17,6 +18,29 @@ const Navbar = () => {
   const accessToken = localStorage.getItem('accessToken');
   const { username, userId } = useSelector((store) => store.user);
   const dispatch = useDispatch();
+
+  useEffect(() => {
+    if (accessToken) {
+      const options = {
+        method: 'GET',
+        headers: {
+          // prettier-ignore
+          'Authorization': accessToken,
+          'Content-Type': 'application/json',
+        },
+      };
+      fetch(API_URL('user'), options)
+        .then((response) => response.json())
+        .then((data) =>
+          batch(() => {
+            dispatch(user.actions.setUsername(data.response.username));
+            dispatch(user.actions.setUserId(data.response.id));
+            dispatch(user.actions.setEmail(data.response.email));
+            dispatch(user.actions.setError(null));
+          })
+        );
+    }
+  }, [accessToken]);
 
   const onLogOut = () => {
     dispatch(user.actions.logout());
@@ -30,28 +54,15 @@ const Navbar = () => {
     <StyledNavbar>
       <NavbarContainer>
         <div>
-          {!accessToken ? (
+          <Link to={`/`}>
             <Img src={Logo} alt="Quizzies logo" />
-          ) : (
-            <Link to={`/`}>
-              <Img src={Logo} alt="Quizzies logo" />
-            </Link>
-          )}
+          </Link>
         </div>
         <Logos>
           <div>
             {accessToken ? (
               <Link to={`/home`}>
                 <AiOutlineHome style={iconStyles} />
-              </Link>
-            ) : (
-              <div></div>
-            )}
-          </div>
-          <div>
-            {accessToken ? (
-              <Link to={`/profile`}>
-                <CgProfile style={iconStyles} />
               </Link>
             ) : (
               <div></div>
@@ -68,17 +79,24 @@ const Navbar = () => {
               </Link>
             )}
           </div>
-          <div>
+          <LinkWrapper>
             {accessToken ? (
-              <GhostBtn type="button" onClick={onLogOut}>
-                <RiLogoutBoxRLine style={iconStyles} />
-              </GhostBtn>
-            ) : (
-              <Link to={`/`}>
-                <AiOutlineHome style={iconStyles} />
+              <Link to={`/profile`}>
+                <ProfileIconWrapper>
+                  <CgProfile style={iconStyles} /> {username && username}
+                  <Menu className="profile-menu">
+                    <GhostBtn type="button" onClick={onLogOut}>
+                      <LinkWrapper>
+                        Logout <RiLogoutBoxRLine style={iconStyles} />
+                      </LinkWrapper>
+                    </GhostBtn>
+                  </Menu>
+                </ProfileIconWrapper>
               </Link>
+            ) : (
+              <div></div>
             )}
-          </div>
+          </LinkWrapper>
         </Logos>
       </NavbarContainer>
     </StyledNavbar>
@@ -92,6 +110,31 @@ const NavbarContainer = styled.div`
   justify-content: space-between;
   align-items: center;
   margin-right: 20px;
+`;
+
+const LinkWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 5px;
+`;
+
+const ProfileIconWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  position: relative;
+  &:hover .profile-menu {
+    display: flex;
+  }
+`;
+
+const Menu = styled.div`
+  display: none;
+  position: absolute;
+  background: #bb5ff7;
+  bottom: -45px;
+  padding: 10px;
+  right: 0;
 `;
 
 const Logos = styled.div`
