+++ frontend/src/components/pages/Profile.js
@@ -1,8 +1,6 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector, batch } from 'react-redux';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import { API_URL } from 'utils/urls';
-import user from 'reducers/auth';
 import styled from 'styled-components/macro';
 import {
   Container,
@@ -15,31 +13,6 @@ import { Link } from 'react-router-dom';
 const Profile = () => {
   const { username } = useSelector((store) => store.user);
   const accessToken = localStorage.getItem('accessToken');
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (accessToken) {
-      const options = {
-        method: 'GET',
-        headers: {
-          // prettier-ignore
-          'Authorization': accessToken,
-          'Content-Type': 'application/json',
-        },
-      };
-      fetch(API_URL('user'), options)
-        .then((response) => response.json())
-        .then((data) =>
-          batch(() => {
-            dispatch(user.actions.setUsername(data.response.username));
-            dispatch(user.actions.setUserId(data.response.id));
-            dispatch(user.actions.setEmail(data.response.email));
-            dispatch(user.actions.setError(null));
-          })
-        );
-    }
-  }, [accessToken]);
 
   if (!accessToken) {
     return <Navigate to="/login" />;
